Tests are passing with new sdk
