+++ angular/package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "server": "npm run build && cd server && node index.js",
+    "server": "npm run build && cd ../demo-server && cross-env BOWTIE_STATIC_CONTENT='../angular/dist/bowtie-sdk-angular-demo' node index.js",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
@@ -20,9 +20,6 @@
     "@angular/platform-browser-dynamic": "~12.0.5",
     "@angular/router": "~12.0.5",
     "@youngalfred/bowtie-sdk": "1.0.6",
-    "axios": "^0.21.1",
-    "form-data": "^4.0.0",
-    "multer": "^1.4.4",
     "rxjs": "~6.6.0",
     "tslib": "^2.1.0",
     "zone.js": "~0.11.4"
@@ -33,7 +30,7 @@
     "@angular/compiler-cli": "~12.0.5",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
-    "express": "^4.17.1",
+    "cross-env": "^7.0.3",
     "jasmine-core": "~3.7.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
