+++ angular/src/app/field-group/field-group.component.css
@@ -8,7 +8,7 @@
     border-radius: 5px;
     display: flex;
     flex-wrap: wrap;
-    background-color: burlywood;
+    max-width: 100%;
 }
 
 .invalid-msg-container {
