+++ angular/src/styles.css
@@ -15,8 +15,7 @@
 
 .flex-wrap {
     display: flex;
-    flex-wrap: wrap;
-        
+    flex-wrap: wrap; 
 }
 
 .flex-wrap > * {
