+++ vanilla-js/package.json
@@ -5,7 +5,7 @@
     "main": "index.js",
     "scripts": {
         "build": "rimraf build/* && webpack --config webpack.config.js && cp -R images build/",
-        "server": "npm run build && node server/server.js",
+        "server": "npm run build && cd ../demo-server && cross-env BOWTIE_STATIC_CONTENT='../vanilla-js/build' node index.js",
         "test": "cd tests && npx wdio wdio.conf.js"
     },
     "repository": {
@@ -18,9 +18,8 @@
         "@babel/core": "^7.12.10",
         "@babel/preset-env": "^7.13.10",
         "babel-loader": "^8.2.2",
+        "cross-env": "^7.0.3",
         "html-webpack-plugin": "^5.3.1",
-        "axios": "^0.21.1",
-        "express": "^4.17.1",
         "rimraf": "^3.0.2",
         "webpack": "^5.25.0",
         "webpack-cli": "^4.5.0"
